Remove unused Vite file system watcher
