+++ .changeset/ninety-otters-repeat.md
@@ -1,0 +1,5 @@
+---
+"@react-router/dev": patch
+---
+
+Remove unused Vite file system watcher
